requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-smorest==0.34.1
PyYAML==6.0.1
timeloop==1.0.2
requests==2.28.1
xgboost==1.7.5
pandas==2.0.1
joblib==1.2.0
scikit-learn==1.2.2
# Dev requirements
black ==22.3.0
pylint ==2.14.3
